finished quotation.php
generate pdf quotation report
calculate quantity prices
add tcpdf library to generate report

+++ html/quotation.php
@@ -18,11 +18,21 @@
 
     if(isset($_GET['remove'])) {
 	    
-        $sql = "DELETE FROM `quotation` WHERE `id` = {$_GET['remove']}";
+        $sql = "DELETE FROM `quotation` WHERE `customer_id` = '{$_SESSION['digimart_current_user_id']}' AND `id` = {$_GET['remove']}";
         
         mysqli_query($conn, $sql);
         
-        header('Location: cart.php');
+        header('Location: quotation.php');
+    
+    }
+
+    if(isset($_GET['removeAll'])) {
+	    
+        $sql = "DELETE FROM `quotation` WHERE `customer_id` = '{$_SESSION['digimart_current_user_id']}'";
+        
+        mysqli_query($conn, $sql);
+        
+        header('Location: quotation.php');
     
     }
 
@@ -92,6 +102,14 @@
             background-color: transparent;
         }
         
+        #getQuoteDisabled {
+            display: block;
+        }
+        
+        #getQuote {
+            display: none;
+        }
+        
     </style>
     
     <script>
@@ -121,6 +139,7 @@
             document.getElementById("qtyPrice<?php echo $row['cartId'] ?>").innerHTML = "LKR " + qtyPrice.toFixed(2);
             document.getElementById("itemTotal<?php echo $row['cartId'] ?>").innerHTML = qtyPrice.toFixed(2);
             document.getElementById("itemTotal1<?php echo $row['cartId'] ?>").value = qtyPrice.toFixed(2);
+            document.getElementById("itemQty<?php echo $row['cartId'] ?>").innerHTML = qty;
             
             calculateTotal();
         }
@@ -131,8 +150,12 @@
             
             var total = 0.0;
             
-            <?php 
-
+            var productId = [];
+            var productQty = [];
+            var productPrice = [];
+            
+            <?php
+            
                 $query2 = "SELECT c.`id` AS 'cartId', p.* FROM `quotation` c, `product` p WHERE c.`product_id` = p.`id` AND c.`customer_id` = '{$_SESSION['digimart_current_user_id']}' ORDER BY `date_time` DESC";
 
                 $result = $conn->query($query2);
@@ -143,21 +166,49 @@
             
             if(document.getElementById("check<?php echo $row['cartId'] ?>").checked == true){
                 total = total + parseFloat(document.getElementById("itemTotal1<?php echo $row['cartId'] ?>").value);
-                document.getElementById("itemTotalLabel<?php echo $row['cartId'] ?>").style.display = "block";
+                document.getElementById("itemId<?php echo $row['cartId'] ?>").style.display = "block";
                 document.getElementById("itemTotal<?php echo $row['cartId'] ?>").style.display = "block";
                 document.getElementById("card-footer<?php echo $row['cartId'] ?>").style.background = "rgba(221,18,60,0.1)";
                 document.getElementById("card-header<?php echo $row['cartId'] ?>").style.background = "rgba(221,18,60,0.1)";
+                
+                var pQty = document.getElementById("qty<?php echo $row['cartId'] ?>").value;
+                var pPrice = document.getElementById("price<?php echo $row['cartId'] ?>").value;
+                
+                productId.push("<?php echo $row['id'] ?>");
+                productQty.push(pQty);
+                productPrice.push(pPrice);
+                
+                
             } else {
-                document.getElementById("itemTotalLabel<?php echo $row['cartId'] ?>").style.display = "none";
+                document.getElementById("itemId<?php echo $row['cartId'] ?>").style.display = "none";
                 document.getElementById("itemTotal<?php echo $row['cartId'] ?>").style.display = "none";
                 document.getElementById("card-footer<?php echo $row['cartId'] ?>").style.background = "rgba(0,0,0,.03)";
                 document.getElementById("card-header<?php echo $row['cartId'] ?>").style.background = "rgba(0,0,0,.03)";
+                
             }
             
             <?php } ?>
             
             document.getElementById("totalPrice").innerHTML = "LKR " + total.toFixed(2);
+            
+            var strLink = "/test/report/tcpdf_lib/examples/quotation_format.php?itemCount=" + productId.length + "&productId=" + productId + "&productPrice=" + productPrice + "&productQty=" + productQty + "&total=" + total;
+            document.getElementById("getQuote").setAttribute("href",strLink);
+            
+            if(total != 0.0) {
+                document.getElementById("getQuote").style.display = "block";
+                document.getElementById("getQuoteDisabled").style.display = "none";
+            } else {
+                document.getElementById("getQuote").style.display = "none";
+                document.getElementById("getQuoteDisabled").style.display = "block";
+            }
+        
         }
+        
+        /*function checkAll() {
+            $("#checkAll").change(function () {
+                $("input:checkbox").prop('checked', $(this).prop("checked"));
+            });
+        }*/
     </script>
     
     
@@ -208,9 +259,16 @@
         
         
         <div class="container">
+            
             <div class="row mt-3">
 
                 <div class="col-lg-8">
+                    
+                    <div class="row justify-content-end mw-100" id="product-container">
+                        <div class="mb-4 d-flex mx-2">
+                            <a href='quotation.php?removeAll=1' onclick="return confirm('This action will remove all item from your quotation.');" class="btn btn-outline-danger w-100 px-5"><i class='far fa-trash-alt fa-lg'></i> Remove All</a>
+                        </div>
+                    </div>
                     
                     <?php 
 
@@ -228,7 +286,7 @@
                                 <div class="card-header d-flex justify-content-between" id="card-header<?php echo $row['cartId'] ?>">
                                     <h6 class="lead">Product Id : <?php echo $row['id']; ?></h6>
                                     <div class="custom-control custom-checkbox">
-                                        <input type="checkbox" class="custom-control-input" id="check<?php echo $row['cartId'] ?>" onclick="qtyPrice<?php echo $row['cartId'] ?>()">
+                                        <input type="checkbox" class="custom-control-input checkItem" id="check<?php echo $row['cartId'] ?>" onclick="qtyPrice<?php echo $row['cartId'] ?>()" onchange="qtyPrice<?php echo $row['cartId'] ?>()">
                                         <label class="custom-control-label" for="check<?php echo $row['cartId'] ?>"></label>
                                     </div>
                                 </div>
@@ -244,7 +302,7 @@
                                         <input type="text" id="price<?php echo $row['cartId'] ?>" value="<?php echo $row['price']; ?>" hidden>
                                         <div class="pl-3 form-group row d-flex justify-content-start">
                                             <label for="qty" class="col-form-label col-form-label-sm">Quentity </label>
-                                            <input type="number" class="form-control form-control-sm col-3 mx-2 <?php if(isset($_COOKIE['theme']) && ($_COOKIE['theme']=='dark'))echo "text-white"; ?>" id="qty<?php echo $row['cartId'] ?>" min="1" value="1" onchange="qtyPrice<?php echo $row['cartId'] ?>()" onkeydown="qtyPrice<?php echo $row['cartId'] ?>()" onkeyup="qtyPrice<?php echo $row['cartId'] ?>()">
+                                            <input type="number" class="form-control form-control-sm col-3 mx-2 <?php if(isset($_COOKIE['theme']) && ($_COOKIE['theme']=='dark'))echo "text-white"; ?>" id="qty<?php echo $row['cartId'] ?>" min="1" value="1" onchange="qtyPrice<?php echo $row['cartId'] ?>()" onkeydown="qtyPrice<?php echo $row['cartId'] ?>()" onkeyup="qtyPrice<?php echo $row['cartId'] ?>()" onselect="qtyPrice<?php echo $row['cartId'] ?>()">
                                             <h5 id="qtyPrice<?php echo $row['cartId'] ?>" class="text-secondary">LKR <?php echo $row['price']; ?></h5>
                                         </div>
                                     </div>
@@ -252,8 +310,8 @@
                                 
                                 <div class="card-footer text-right" id="card-footer<?php echo $row['cartId'] ?>">
                                     <?php
-                                        echo "<a href='cart.php?remove={$row['cartId']}' class='text-danger mx-5'><i class='far fa-trash-alt fa-lg'></i></a>";
-                                        echo "<a href='cart.php' class='btn btn-outline-danger px-5'>Buy</a>";
+                                        echo "<a href='quotation.php?remove={$row['cartId']}' onclick=\"return confirm('This action will remove this item from your quotation.');\" class='text-danger mx-5'><i class='far fa-trash-alt fa-lg'></i></a>";
+                                        echo "<a href='quotation.php' class='btn btn-outline-danger px-5'>Buy</a>";
                                     ?>
                                 </div>
                             </div>
@@ -270,8 +328,7 @@
                     <h2>Quotation</h2>
                     
                     <div class="d-flex justify-content-between">
-                        <div class="p-2 font-weight-bold">PRODUCT ID</div>
-                        <div class="p-2 font-weight-bold">Qty</div>
+                        <div class="p-2 font-weight-bold">PRODUCT ID X QTY</div>
                         <div class="p-2 font-weight-bold">PRICE (LKR)</div>
                     </div>
                     
@@ -286,7 +343,7 @@
                     ?>
                     
                     <div class="d-flex justify-content-between">
-                        <div class="p-2" id="itemTotalLabel<?php echo $row['cartId'] ?>"><?php echo $row['cartId'] ?></div>
+                        <div class="p-2" id="itemId<?php echo $row['cartId'] ?>"><?php echo $row['id'] ?> X <font id="itemQty<?php echo $row['cartId'] ?>">1</font></div>
                         <div class="p-2" id="itemTotal<?php echo $row['cartId'] ?>"><?php echo $row['price']; ?></div>
                         <input type="text" id="itemTotal1<?php echo $row['cartId'] ?>" value="<?php echo $row['price']; ?>" hidden>
                     </div>
@@ -298,7 +355,8 @@
                         <div class="p-2"><h5 id="totalPrice"></h5></div>
                     </div>
                     
-                    <button type="submit" name="msgSend" class="btn btn-danger w-100 mt-3">Get Quotation as PDF</button>
+                    <button id="getQuoteDisabled" class="btn btn-outline-danger w-100 mt-3" data-toggle="tooltip" data-placement="bottom" title="Select item" disabled>Get Quotation as PDF</button>
+                    <a href='' id="getQuote" target="_blank" name="getQuote" class="btn btn-danger w-100 mt-3">Get Quotation as PDF</a>
 
                 </div>
                 <!-- /.col-lg-3 -->
